requirements.txt
numpy>=1.26.0
pyyaml>=6.0.2
python-dotenv>=1.1.0
pydantic>=2.11.0
aiohttp>=3.12.0
datasets>=2.14.0
pillow>=10.0.0
openai>=1.0.0
anthropic>=0.30.0
tqdm>=4.66.0

